backend/requirements.txt
# Core dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.2
pydantic-settings==2.1.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
email-validator==2.1.0
httpx==0.25.2

# Database
sqlalchemy==2.0.23
asyncpg==0.29.0
alembic==1.13.0
pgvector==0.2.4
psycopg2-binary==2.9.9
neo4j==5.28.1

# Redis & Celery
redis==5.0.1
celery==5.3.4

# Document processing
pypdf==3.17.1
python-docx==1.1.0
beautifulsoup4==4.12.2
chardet==5.2.0
python-magic==0.4.27
markdown==3.5.1

# AI/ML
openai==1.52.0
tiktoken==0.5.2
scikit-learn==1.3.2
numpy==1.26.2

# Storage
minio==7.2.0
boto3==1.34.4

# Monitoring & Logging
prometheus-client==0.19.0
python-json-logger==2.0.7

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0

# Other utilities
pyyaml==6.0.1
python-dateutil==2.8.2
tenacity==8.2.3
aiofiles==23.2.1

# Development
black==23.12.0
flake8==6.1.0
mypy==1.7.1 